+++ erudit/conf/settings.py
@@ -42,6 +42,8 @@ DEFAULT_JOURNAL_PROVIDERS = {
             'collection_title': 'Persée',
             'collection_code': 'persee',
             'endpoint': 'http://oai.persee.fr/oai',
+            'issue_metadataprefix': 'persee_mets',
+            'article_metadataprefix': 'oai_dc',
         },
     ],
 }